add flag checking for pruning waypoints
